refactoring(15): split resultSections leaf layer into resultTypes + resultHelpers
Per refactoring/notes/03 DAG: move the shared types (OrsRenderInput,
RubricNode, LeafGrade, StageState) into resultTypes.ts and the pure
helpers / string formatters (tryParseJson, buildGradeMap, aggregateScore,
truncSha, fileCount, asFileList, formatFiles, format*Stage, *State) into
resultHelpers.ts. Byte-exact verbatim slices (modulo export/import);
resultSections.tsx imports them back. 1601->1417 lines. No behavior change.
